+++ dropwizard-guicey/src/test/groovy/ru/vyarus/dropwizard/guice/config/sharedstate/SharedStateTest.groovy
@@ -101,7 +101,8 @@ class SharedStateTest extends Specification {
         when: "get with null supplier"
         def res = state.get(App, null)
         then: "behave as usual get"
-        res == null
+        def exN = thrown(NullPointerException)
+        exN.message == "Cannot invoke \"java.util.function.Supplier.get()\" because \"defaultValue\" is null"
 
         when: "duplicate assign"
         state.assignTo app