+++ CHANGELOG.md
@@ -40,7 +40,9 @@
       Add .whenSharedStateReady() for GuiceyBootstrap and GuiceyEnvironment
       (not required for the main bundle as there is withSharedState() method where whenReady() could be used directly)
     - (breaking) Tie a state key to the stored object type to simplify usage (type-safe) and force
-       state objects usage instead of whatever values 
+       state objects usage instead of whatever values
+    - Fix null value supplier behavior (not allowed): .get(key, supplier)
+    - Add shared state usage report: .printSharedStateUsage()
 * Transitive guicey bundles initialize immediately after registration (unify behavior with dropwizard bundles and guice modules) 
 * Test improvements:
   - Track guicey test extensions time (would appear when debug option enabled)